Ubah lagi plis

+++ database/migrations/2025_12_07_065922_add_role_to_users_table.php
@@ -11,14 +11,18 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('role')->default('user'); // default semua user biasa
-        });
+            if(!Schema::hasColumn('users', 'role')) {
+              $table->string('role')->default('user'); // default semua user biasa
+            }
+         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if(Schema::hasColumn('users', 'role')) {
+               $table->dropColumn('role');
+            }
         });
     }
